Create README.md
As a self-motivated and detail-oriented professional, I am able to work independently as well as part of a team. I pride myself on my ability to troubleshoot and solve complex problems in a fast-paced environment, and I am always looking for new and innovative ways to improve my skills and stay up-to-date with the latest industry trends.